refactor method names for tokenizer trait and lexer
makes things a bit clearer, dontcha think? previously the `scan_token`
method was doing a bit too much - advancing the state, matching the
token type, and adding the token. moving all that to the tokenize method
and renaming scan to next and only having it get the next token makes
more logical sense